Move builders into a different class
