+++ internal/jennies/java/templates/builders/assigments.tmpl
