+++ internal/jennies/java/types.go
@@ -148,7 +148,7 @@ func (tf *typeFormatter) defaultValueFor(def ast.Type) string {
 	case ast.KindRef:
 		refDef := fmt.Sprintf("%s.%s", def.AsRef().ReferredPkg, def.AsRef().ReferredType)
 		if tf.typeHasBuilder(def) {
-			return fmt.Sprintf("new %s.Builder().build()", tf.formatPackage(refDef))
+			return fmt.Sprintf("new %sBuilder().build()", tf.formatPackage(refDef))
 		}
 		return fmt.Sprintf("new %s()", tf.formatPackage(refDef))
 	case ast.KindStruct:
